assign tasks considering data access levels

+++ pybossa/view/account.py
@@ -62,7 +62,9 @@ from pybossa import otp
 import time
 from pybossa.cache.users import get_user_preferences
 from pybossa.sched import release_user_locks
-from pybossa.core import private_instance_params
+from pybossa.data_access import (data_access_levels, ensure_data_access_assignment_from_form,
+    copy_data_access_levels)
+import app_settings
 
 blueprint = Blueprint('account', __name__)
 
@@ -330,7 +332,7 @@ def register():
     """
     if current_app.config.get('LDAP_HOST', False):
         return abort(404)
-    if not current_app.config.upref_mdata:
+    if not app_settings.upref_mdata:
         form = RegisterForm(request.body)
     else:
         form = RegisterFormWithUserPrefMetadata(request.body)
@@ -340,7 +342,7 @@ def register():
     msg = "I accept receiving emails from %s" % current_app.config.get('BRAND')
     form.consent.label = msg
     if request.method == 'POST' and form.validate():
-        if current_app.config.upref_mdata:
+        if app_settings.upref_mdata:
             user_pref, metadata = get_user_pref_and_metadata(form.name.data, form)
             account = dict(fullname=form.fullname.data, name=form.name.data,
                            email_addr=form.email_addr.data,
@@ -354,8 +356,7 @@ def register():
                            email_addr=form.email_addr.data,
                            password=form.password.data,
                            consent=form.consent.data)
-        if private_instance_params:
-            account['data_access'] = form.data_access.data
+        ensure_data_access_assignment_from_form(account, form)
         confirm_url = get_email_confirmation_url(account)
         if current_app.config.get('ACCOUNT_CONFIRMATION_DISABLED'):
             project_slugs=form.project_slug.data
@@ -448,8 +449,9 @@ def create_account(user_data, project_slugs=None, ldap_disabled=True):
     else:
         if user_data.get('ldap'):
             new_user.ldap = user_data['ldap']
-    if private_instance_params:
-        new_user.info['data_access'] = user_data.get('data_access', [])
+
+    if 'info' in user_data:
+        copy_data_access_levels(new_user.info, user_data)
     user_repo.save(new_user)
     if not ldap_disabled:
         flash(gettext('Thanks for signing-up'), 'success')
@@ -479,7 +481,7 @@ def redirect_profile():
         return redirect_content_type(url_for('.signin'), status='not_signed_in')
     if (request.headers.get('Content-Type') == 'application/json') and current_user.is_authenticated():
         form = None
-        if current_app.config.upref_mdata:
+        if app_settings.upref_mdata:
             form_data = cached_users.get_user_pref_metadata(current_user.name)
             form = UserPrefMetadataForm(**form_data)
             form.set_upref_mdata_choices()
@@ -502,7 +504,7 @@ def profile(name):
         raise abort(404)
 
     form = None
-    if current_app.config.upref_mdata:
+    if app_settings.upref_mdata:
         form_data = cached_users.get_user_pref_metadata(user.name)
         form = UserPrefMetadataForm(**form_data)
         form.set_upref_mdata_choices()
@@ -540,7 +542,8 @@ def _show_public_profile(user, form, can_update):
                     percentage_tasks_completed=percentage_tasks_completed,
                     form=form,
                     can_update=can_update,
-                    private_instance=bool(private_instance_params))
+                    private_instance=bool(data_access_levels),
+                    upref_mdata_enabled=bool(app_settings.upref_mdata))
 
     return handle_content_type(response)
 
@@ -563,7 +566,8 @@ def _show_own_profile(user, form, can_update):
                     user=user_dict,
                     form=form,
                     can_update=can_update,
-                    private_instance=bool(private_instance_params))
+                    private_instance=bool(data_access_levels),
+                    upref_mdata_enabled=bool(app_settings.upref_mdata))
 
     return handle_content_type(response)
 
@@ -1035,15 +1039,16 @@ def add_metadata(name):
                                percentage_tasks_completed=percentage_tasks_completed,
                                form=form,
                                input_form=True,
-                               can_update=can_update)
+                               can_update=can_update,
+                               upref_mdata_enabled=bool(app_settings.upref_mdata))
 
     user_pref, metadata = get_user_pref_and_metadata(name, form)
     user.info['metadata'] = metadata
-    if private_instance_params:
-        user.info['data_access'] = form.data_access.data
+    ensure_data_access_assignment_from_form(user.info, form)
     user.user_pref = user_pref
     user_repo.update(user)
     cached_users.delete_user_pref_metadata(user.name)
+    cached_users.delete_user_access_levels_by_id(user.id)
     delete_memoized(get_user_preferences, user.id)
     flash("Input saved successfully", "info")
     return redirect(url_for('account.profile', name=name))
